// frontend/src/types/api/listApprovalsApiApprovalsGetSort.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListApprovalsApiApprovalsGetSort = typeof ListApprovalsApiApprovalsGetSort[keyof typeof ListApprovalsApiApprovalsGetSort];


export const ListApprovalsApiApprovalsGetSort = {
  newest: 'newest',
  oldest: 'oldest',
} as const;